Trait compile::ast::AstExpr

pub trait AstExpr {
    fn offset_spans(&mut self, offset: usize);
}

Common functionality for AST expression nodes

Required methods

fn offset_spans(&mut self, offset: usize)

Increase the indices of the Spans in this AST node by a given offset. This is useful when this AST node is part of a format string, and the indices need to be offset to be accurate for the entire file.